Updates to make safer for multiple routers, select routers by hostname.
This adds hostname labels to the targets in Prometheus, and then drives
the data selection using that vs. the IP addresses in the 'instance'
label.  Also, cleans up some minor SNMP config, and allows passing
a 'mib' label per-target in case you're trying to migrate a bunch
of routers from the original MIB to the enhanced one (e.g. by cloning
one to `ddwrt_legacy` so you can uncomment the legacy OIDs and comment
out the new OIDs in that instance).

Signed-off-by: Rafal Boni <rafal@yesware.com>